# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reikun API",
    "description": "Japanese dictionary search (hybrid dense + BM25), kanji details and JLPT-calibrated grammar explanations.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search Endpoint",
        "description": "Same pipeline as the dictionary UI: rewrite → route → retrieve.\nResults carry `ruby` parts for furigana rendering; Japanese sentences that\ndecompose into word units come back in `segments` (the \"parsed as\" chips).",
        "operationId": "search_endpoint_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "pattern": "\\S",
              "description": "English or Japanese word, or a natural-language question",
              "title": "Q"
            },
            "description": "English or Japanese word, or a natural-language question"
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Number of results",
              "default": 10,
              "title": "N"
            },
            "description": "Number of results"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "auto",
                "hybrid",
                "vector",
                "text"
              ],
              "type": "string",
              "description": "auto routes the query to the best plan; the rest are fixed pipelines",
              "default": "auto",
              "title": "Mode"
            },
            "description": "auto routes the query to the best plan; the rest are fixed pipelines"
          },
          {
            "name": "rewrite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Normalise natural-language queries to a dictionary gloss first",
              "default": true,
              "title": "Rewrite"
            },
            "description": "Normalise natural-language queries to a dictionary gloss first"
          },
          {
            "name": "log",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Record the search in the telemetry sink",
              "default": true,
              "title": "Log"
            },
            "description": "Record the search in the telemetry sink"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kanji/random": {
      "get": {
        "tags": [
          "kanji"
        ],
        "summary": "Kanji Random Endpoint",
        "description": "A random kanji that is itself a JMdict headword — the shuffle button's\npick. Same rules as the UI: frequency-ranked pool, optionally restricted\nto a JLPT level, up to 8 headword-verified tries then an unchecked pick.",
        "operationId": "kanji_random_endpoint_kanji_random_get",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "N5",
                    "N4",
                    "N3",
                    "N2",
                    "N1"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Prefer kanji tagged at this JLPT level (falls back to the common pool)",
              "title": "Level"
            },
            "description": "Prefer kanji tagged at this JLPT level (falls back to the common pool)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RandomKanji"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kanji": {
      "get": {
        "tags": [
          "kanji"
        ],
        "summary": "Kanji Batch Endpoint",
        "description": "Compact hover-card data for several kanji at once — one round-trip per\nsentence of tooltips. Unknown kanji map to null.",
        "operationId": "kanji_batch_endpoint_kanji_get",
        "parameters": [
          {
            "name": "chars",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "Kanji characters to look up (deduped; every char is treated as one kanji)",
              "title": "Chars"
            },
            "description": "Kanji characters to look up (deduped; every char is treated as one kanji)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/KanjiHover"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "title": "Response Kanji Batch Endpoint Kanji Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kanji/{char}": {
      "get": {
        "tags": [
          "kanji"
        ],
        "summary": "Kanji Endpoint",
        "description": "KANJIDIC2 details (meta badges, readings with okurigana parts, common\nwords with ruby parts) for one kanji — the frontend renders the card.",
        "operationId": "kanji_endpoint_kanji__char__get",
        "parameters": [
          {
            "name": "char",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Char"
            }
          },
          {
            "name": "strokes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include KanjiVG stroke order as structured data (viewBox, path d's, number labels)",
              "default": false,
              "title": "Strokes"
            },
            "description": "Include KanjiVG stroke order as structured data (viewBox, path d's, number labels)"
          },
          {
            "name": "log",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Log"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KanjiCard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/explain": {
      "post": {
        "tags": [
          "grammar"
        ],
        "summary": "Explain Endpoint",
        "description": "JLPT-level-calibrated grammar explanation (Groq). Markdown bullets.",
        "operationId": "explain_endpoint_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/explain/stream": {
      "post": {
        "tags": [
          "grammar"
        ],
        "summary": "Explain Stream Endpoint",
        "description": "Server-sent events: `data: {\"text\": \"…\"}` chunks as Groq streams them,\nthen `data: {\"done\": true, \"explanation_id\": …, \"latency_ms\": …}`.",
        "operationId": "explain_stream_endpoint_explain_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Feedback Endpoint",
        "description": "Thumbs up (+1) / down (-1) on a search or an explanation.",
        "operationId": "feedback_endpoint_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExplainRequest": {
        "properties": {
          "sentence": {
            "type": "string",
            "minLength": 1,
            "title": "Sentence",
            "description": "Japanese example sentence"
          },
          "english": {
            "type": "string",
            "title": "English",
            "description": "English translation (improves the explanation)",
            "default": ""
          },
          "level": {
            "type": "string",
            "enum": [
              "N5",
              "N4",
              "N3",
              "N2",
              "N1"
            ],
            "title": "Level",
            "default": "N5"
          }
        },
        "type": "object",
        "required": [
          "sentence"
        ],
        "title": "ExplainRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KanjiCard": {
        "properties": {
          "literal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Literal"
          },
          "codepoint_hex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codepoint Hex"
          },
          "stroke_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stroke Count"
          },
          "grade": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "freq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freq"
          },
          "jlpt_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jlpt Level"
          },
          "meanings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Meanings",
            "default": []
          },
          "meta": {
            "items": {
              "$ref": "#/components/schemas/MetaBadge"
            },
            "type": "array",
            "title": "Meta",
            "default": []
          },
          "on_yomi": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "On Yomi",
            "default": []
          },
          "kun_yomi": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kun Yomi",
            "default": []
          },
          "readings": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ReadingChip"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Readings",
            "default": {}
          },
          "common_words": {
            "items": {
              "$ref": "#/components/schemas/CommonWord"
            },
            "type": "array",
            "title": "Common Words",
            "default": []
          },
          "stroke": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StrokeData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "KanjiCard"
      },
      "KanjiHover": {
        "properties": {
          "literal": {
            "type": "string",
            "title": "Literal"
          },
          "meanings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Meanings",
            "default": []
          },
          "on_yomi": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "On Yomi",
            "default": []
          },
          "kun_yomi": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kun Yomi",
            "default": []
          },
          "meta": {
            "items": {
              "$ref": "#/components/schemas/MetaBadge"
            },
            "type": "array",
            "title": "Meta",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "literal"
        ],
        "title": "KanjiHover",
        "description": "Compact card for kanji-in-text hover tooltips (batch endpoint)."
      },
...
      "RandomKanji": {
        "properties": {
          "kanji": {
            "type": "string",
            "title": "Kanji"
          }
        },
        "type": "object",
        "required": [
          "kanji"
        ],
        "title": "RandomKanji"
      },
...
      "SearchResponseModel": {
        "properties": {
          "search_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "rewrite": {
            "$ref": "#/components/schemas/RewriteInfo"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "segments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Segment"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          },
          "meta": {
            "$ref": "#/components/schemas/SearchMeta",
            "default": {
              "cached": false
            }
          }
        },
        "type": "object",
        "required": [
          "results",
          "rewrite",
          "mode",
          "latency_ms"
        ],
        "title": "SearchResponseModel"
      },
...
    }
  }
}